% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoenesSkript.R
\name{firstfunction}
\alias{firstfunction}
\title{Irgendeinen Satz über R schreiben}
\usage{
firstfunction(x)
}
\arguments{
\item{x}{A number}
}
\value{
Er schreibt einfach einen recht sinnlosen Satz.
}
\description{
Irgendeinen Satz über R schreiben
}
\examples{
firstfunction(1, 1)
add(10, 1)

}
